fix UpdateNotify :(
at some point in recent upgrades i broke the notification that you can
refresh and update. All this means is that the user is stuck on the old
version until they either restart their phone or the phone decides to
kill the in-memory version of the app (i think, not 100%).

I was able to get the app to update after hard-killing it on android,
and restarting.